requirements.txt
Flask
psycopg2-binary
PyYAML
pytest
bandit